获取社区详情接口

+++ internal/entity/community.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Community struct {
-	Cid       uint   `gorm:"primarykey"`
+	Cid       int64  `gorm:"primarykey"`
 	Name      string `gorm:"type:varchar(32);unique;not null" json:"name"`
 	Intro     string `gorm:"type:varchar(32);unique;not null" json:"intro"`
 	CreatedAt time.Time
